DESIGN_NOTES: Think some more about what to hash
Chiefly, we want to hash shared-object-exports, so we can compare
captured-back-at-build-time shared-object-exports hashes to actual
available-now shared-object-exports exports in builddep `AbiHash` structs
when doing rebuild analysis.

shared-object-exports hashes are a bit like build-ids, but only for
exported .dynsym symbols, that is, symbols that match whatever
characteristics we end up arguing constitutes public ABI and then write
logic to express and capture in the associated `AbiCapture` exports.

Signed-off-by: Rune Morling <ermo@serpentos.com>